get userid when adding item to shopping list
